MCQ Generator APP/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
PyMuPDF
google-generativeai
python-dotenv